Add an integration test that checks that builds from github JSON
Also
- added a feature to skip tests.
- refactored the test runner

This new integration test needed the skip feature it fails. There is an infinite
recursion when the project file is loaded that needs to be investigated.

+++ test.sh
@@ -7,27 +7,48 @@ test() {
     for d in $(ls $fixture_dir)
     do
         local test_dir=${fixture_dir}/${d}
-        [[ -d $test_dir ]] && {
-            # setup
-            echo "test project: ${test_dir}"
-            pushd $test_dir > /dev/null
-            _prepare_nix_dir
-            _add_current_project_to_nix
-            ./setup_test.sh
+        [[ -e $test_dir ]] || continue
 
-            # test
-            trap 'popd > /dev/null' INT TERM EXIT
-            nix-build --no-out-link
+        pushd $test_dir > /dev/null
+        # maybe skip
+        [[ -f ./SKIP ]] && {
+            local reason=$(cat ./SKIP)
+            local cause="${reason:-'incomplete test'}"
             echo
-            echo "OK: test project : ${test_dir}"
+            echo "SKIPPED: $test_dir: $cause"
             echo
-
-            # cleanup
-            [[ -d nix ]] && rm -fR nix
             popd > /dev/null
-            trap - INT TERM EXIT
+            continue
         }
+        _test_one_project $test_dir
     done
+}
+
+_test_one_project() {
+    local test_dir=${1-''}
+    [[ -z $test_dir ]] && return 1;
+
+    # setup
+    trap 'popd > /dev/null' INT TERM EXIT
+
+    echo
+    echo "testing project: ${test_dir}"
+    echo
+
+    _prepare_nix_dir
+    _add_current_project_to_nix
+    ./setup_test.sh
+
+    # test
+    nix-build --no-out-link
+    echo
+    echo "OK: test project : ${test_dir}"
+    echo
+
+    # cleanup
+    [[ -d nix ]] && rm -fR nix
+    popd > /dev/null
+    trap - INT TERM EXIT
 }
 
 _prepare_nix_dir() {
@@ -47,7 +68,7 @@ _add_current_project_to_nix() {
     mkdir -p nix/nix-expr
     local nix_file="nix/nix-expr/${cwd##*/}.nix"
     cabal2nix . > $nix_file
-    sed -i'' -e 's|src = ./.|src = ../../.|' $nix_file
+    sed -i '' -e 's|src = ./.|src = ../../.|' $nix_file
 }
 
 test
